requirements.txt
-i https://pypi.org/simple
-e .
-e .
aiodns==2.0.0
aiohttp==3.6.3; python_full_version >= '3.5.3'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
bcrypt==3.2.0; python_version >= '3.6'
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fastapi==0.61.1; python_version >= '3.6'
h11==0.11.0
httpcore==0.12.0; python_version >= '3.6'
httpx==0.16.1; python_version >= '3.6'
idna==2.10
multidict==4.7.6; python_version >= '3.5'
passlib==1.7.4
pycares==3.1.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.6.1; python_version >= '3.6'
pyjwt==1.7.1
python-dotenv==0.14.0
rfc3986[idna2008]==1.4.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.13.6; python_version >= '3.6'
uvicorn==0.12.2
yarl==1.5.1
